Filtering Parameters
User can update the filtering parameters for each separation. Their settings are stored in a database but that information is
currently hidden from the user (so they can't check it and compare it later on).

Filtering is performed after the user selects the separations. In order to see the change the user must re-select the separations of interest.