Peak and Peak LUT for sql database.
Relationship is Peak LUT holds general information about the peak, and Peak holds specivic information of that peak with a connection to a given separation.id
Updates for the Display so that the peaks LUT has a datatable that can be editable.

Curious about how to best connect peak look ups to separation. I like the idea of using a buffer, but it could also be a tag. That way you could sort by tags. I need to remove the contstraint that the name is unique. There may need to be many different samples but that are differentiated by the buffers that are used.